+++ Logger.h
@@ -1,0 +1,13 @@
+#pragma once
+#include "Sphere.h"
+#include <string>
+using namespace std;
+class Logger
+{
+public:
+    string log; 
+
+public:
+    void addRecord(Sphere& sphere);
+    void saveLog();
+};
