+++ Sphere.h
@@ -1,10 +1,11 @@
+#pragma once
 #include <string>
 
 using namespace std;
 
-class Sphere {
-
-    static int instanceCount;
+class Sphere
+{
+static int instanceCount;
 
 public:
     string objectName; 
@@ -12,13 +13,5 @@ public:
     
     Sphere(double r, string name) : radius(r), objectName(name) {instanceCount++;}
     static int getInstanceCount() {return instanceCount;}
-};
-
-class Logger {
     
-    std::string log; 
-
-public:
-    void addRecord(Sphere& sphere);
-    void saveLog();
 };
